requirements.txt
fastapi
uvicorn
langchain
chromadb
psycopg2-binary==2.9.3
python-dotenv
pydantic
huggingface-hub
transformers
flask
requests
torch
accelerate
openai
dotenv
tiktoken
langchain-community
langchain-openai
sqlparse
gradio
plotly
alembic
passlib[bcrypt]
pytest
email-validator
python-jose[cryptography]